refactor(renderer): split WorktreeCard
