+++ src/renderer/src/components/sidebar/use-worktree-card-secondary-details.ts
@@ -1,0 +1,224 @@
+import React, { useCallback } from 'react'
+
+import type { GitHubWorkItem } from '../../../../shared/types'
+import { useAppStore } from '@/store'
+import { hasWorktreeCardDetails } from './WorktreeCardMeta'
+import { usePromptCacheCountdownStartedAt } from './CacheTimer'
+import { useWorktreeAgentRows } from './useWorktreeAgentRows'
+import type { WorktreeCardProps } from './worktree-card-model'
+import type { useWorktreeCardFoundation } from './use-worktree-card-foundation'
+import type { useWorktreeCardLinkedDetails } from './use-worktree-card-linked-details'
+import type { useWorktreeCardReviewDetails } from './use-worktree-card-review-details'
+
+type Foundation = ReturnType<typeof useWorktreeCardFoundation>
+type LinkedDetails = ReturnType<typeof useWorktreeCardLinkedDetails>
+type ReviewDetails = ReturnType<typeof useWorktreeCardReviewDetails>
+
+export function useWorktreeCardSecondaryDetails({
+  worktree,
+  repo,
+  statusPrDisplay,
+  showStatus,
+  showIssue,
+  showLinearIssue,
+  showJiraIssue,
+  showPR,
+  showAutomation,
+  showCli,
+  showComment,
+  showPorts,
+  issueDisplay,
+  linearIssue,
+  linearIssueDisplay,
+  jiraIssueDisplay,
+  prDisplay,
+  linkedGitLabMR,
+  linkedBitbucketPR,
+  linkedAzureDevOpsPR,
+  linkedGiteaPR,
+  cardProps,
+  newCardStyle,
+  compactCards,
+  agentActivityDisplayMode,
+  workspacePorts,
+  openTaskPage,
+  updateWorktreeMeta
+}: Pick<WorktreeCardProps, 'worktree' | 'repo' | 'statusPrDisplay'> &
+  Pick<
+    Foundation,
+    | 'cardProps'
+    | 'newCardStyle'
+    | 'compactCards'
+    | 'agentActivityDisplayMode'
+    | 'workspacePorts'
+    | 'openTaskPage'
+    | 'updateWorktreeMeta'
+  > &
+  Pick<LinkedDetails, 'issueDisplay' | 'linearIssue' | 'linearIssueDisplay' | 'jiraIssueDisplay'> &
+  Pick<
+    ReviewDetails,
+    'prDisplay' | 'linkedGitLabMR' | 'linkedBitbucketPR' | 'linkedAzureDevOpsPR' | 'linkedGiteaPR'
+  > & {
+    showStatus: boolean
+    showIssue: boolean
+    showLinearIssue: boolean
+    showJiraIssue: boolean
+    showPR: boolean
+    showAutomation: boolean
+    showCli: boolean
+    showComment: boolean
+    showPorts: boolean
+  }) {
+  // Why: unread lives in the left status lane, so the Status toggle owns both the dot/PR slot and unread emphasis.
+  const showUnreadEmphasis = showStatus && worktree.isUnread
+  const hoverIssue = issueDisplay
+  const hoverLinearIssue = linearIssueDisplay
+  const hoverJiraIssue = jiraIssueDisplay
+  const hoverReview = prDisplay
+  const statusLaneReview = statusPrDisplay ?? hoverReview
+  const hoverComment = worktree.comment
+  const metaIssue = showIssue ? hoverIssue : null
+  const metaLinearIssue = showLinearIssue ? hoverLinearIssue : null
+  const metaJiraIssue = showJiraIssue ? hoverJiraIssue : null
+  const metaReview = showPR ? hoverReview : null
+  const metaAutomationProvenance = showAutomation ? worktree.automationProvenance : null
+  const metaCliProvenance = showCli ? worktree.cliProvenance : null
+  const metaComment = showComment ? hoverComment : null
+  const showInlineAgentList = cardProps.includes('inline-agents') && (newCardStyle || !compactCards)
+  const compactInlineAgentRows = useWorktreeAgentRows(
+    worktree.id,
+    showInlineAgentList && agentActivityDisplayMode === 'compact'
+  )
+  const compactInlineAgentRowsVisible =
+    showInlineAgentList &&
+    agentActivityDisplayMode === 'compact' &&
+    compactInlineAgentRows.length > 0
+  const showAggregateCacheTimer = !compactCards && !compactInlineAgentRowsVisible
+  const handleOpenGitHubIssueInOrca = useCallback(
+    (e: React.MouseEvent) => {
+      e.stopPropagation()
+      const issueUrl = hoverIssue && 'url' in hoverIssue ? hoverIssue.url : undefined
+      if (!repo || !hoverIssue || !issueUrl) {
+        return
+      }
+      const item: GitHubWorkItem = {
+        id: issueUrl,
+        type: 'issue',
+        number: hoverIssue.number,
+        title: hoverIssue.title,
+        state: 'state' in hoverIssue ? (hoverIssue.state ?? 'open') : 'open',
+        url: issueUrl,
+        labels: 'labels' in hoverIssue ? (hoverIssue.labels ?? []) : [],
+        updatedAt: new Date().toISOString(),
+        author: null,
+        repoId: repo.id
+      }
+      openTaskPage({ taskSource: 'github', preselectedRepoId: repo.id, openGitHubWorkItem: item })
+    },
+    [hoverIssue, openTaskPage, repo]
+  )
+  const handleOpenReviewInOrca = useCallback(
+    (e: React.MouseEvent) => {
+      e.stopPropagation()
+      if (!repo || !hoverReview?.url || hoverReview.provider !== 'github') {
+        return
+      }
+      const item: GitHubWorkItem = {
+        id: hoverReview.url,
+        type: 'pr',
+        number: hoverReview.number,
+        title: hoverReview.title,
+        state: hoverReview.state ?? 'open',
+        url: hoverReview.url,
+        labels: [],
+        updatedAt: 'updatedAt' in hoverReview ? hoverReview.updatedAt : new Date().toISOString(),
+        author: null,
+        headSha: 'headSha' in hoverReview ? hoverReview.headSha : undefined,
+        repoId: repo.id
+      }
+      openTaskPage({ taskSource: 'github', preselectedRepoId: repo.id, openGitHubWorkItem: item })
+    },
+    [hoverReview, openTaskPage, repo]
+  )
+  const hoverReviewProvider = hoverReview?.provider
+  const hasExplicitLinkedReview =
+    (hoverReviewProvider === 'github' && worktree.linkedPR !== null) ||
+    (hoverReviewProvider === 'gitlab' && linkedGitLabMR !== null) ||
+    (hoverReviewProvider === 'bitbucket' && linkedBitbucketPR !== null) ||
+    (hoverReviewProvider === 'azure-devops' && linkedAzureDevOpsPR !== null) ||
+    (hoverReviewProvider === 'gitea' && linkedGiteaPR !== null)
+  const handleUnlinkReview = useCallback(() => {
+    switch (hoverReviewProvider) {
+      case 'github':
+        void updateWorktreeMeta(worktree.id, { linkedPR: null })
+        return
+      case 'gitlab':
+        void updateWorktreeMeta(worktree.id, { linkedGitLabMR: null })
+        return
+      case 'bitbucket':
+        void updateWorktreeMeta(worktree.id, { linkedBitbucketPR: null })
+        return
+      case 'azure-devops':
+        void updateWorktreeMeta(worktree.id, { linkedAzureDevOpsPR: null })
+        return
+      case 'gitea':
+        void updateWorktreeMeta(worktree.id, { linkedGiteaPR: null })
+        break
+      case 'unsupported':
+      case undefined:
+        break
+    }
+  }, [hoverReviewProvider, updateWorktreeMeta, worktree.id])
+  const handleOpenLinearIssueInOrca = useCallback(
+    (e: React.MouseEvent) => {
+      e.stopPropagation()
+      if (!linearIssue) {
+        return
+      }
+      openTaskPage({ taskSource: 'linear', openLinearIssue: linearIssue })
+    },
+    [linearIssue, openTaskPage]
+  )
+  const hasDetails = hasWorktreeCardDetails({
+    issue: metaIssue,
+    linearIssue: metaLinearIssue,
+    jiraIssue: metaJiraIssue,
+    review: newCardStyle ? null : metaReview,
+    comment: metaComment,
+    automationProvenance: metaAutomationProvenance,
+    cliProvenance: metaCliProvenance
+  })
+  const hasPorts = showPorts && workspacePorts.length > 0
+  const cacheStartedAt = usePromptCacheCountdownStartedAt(worktree.id, showAggregateCacheTimer)
+  const cacheTtlMs = useAppStore((s) =>
+    showAggregateCacheTimer ? (s.settings?.promptCacheTtlMs ?? 0) : 0
+  )
+
+  return {
+    showUnreadEmphasis,
+    hoverIssue,
+    hoverLinearIssue,
+    hoverJiraIssue,
+    hoverReview,
+    statusLaneReview,
+    hoverComment,
+    metaIssue,
+    metaLinearIssue,
+    metaJiraIssue,
+    metaReview,
+    metaAutomationProvenance,
+    metaCliProvenance,
+    metaComment,
+    showInlineAgentList,
+    compactInlineAgentRows,
+    handleOpenGitHubIssueInOrca,
+    handleOpenReviewInOrca,
+    hasExplicitLinkedReview,
+    handleUnlinkReview,
+    handleOpenLinearIssueInOrca,
+    hasDetails,
+    hasPorts,
+    cacheStartedAt,
+    cacheTtlMs
+  }
+}
